Implemented random comic download.
It first checks for the actual number of comics and then downloads
a random comic in that range. Since it checks the url of a comic to
see if it exists it may be a bit slow. This is partially avoided by
having the file xkcd_max.txt, which stores the last max retrieved so
that the count starts there.